Add heading to UserLocation and expose UserLocation type (#522)
* Show `UserLocation` type

Objects of this type are returned by framework methods, but client code
and not refer to the type currently.

* Add heading to UserLocation

While UserLocation includes a "bearing" (also called "course")
indicating the direction of travel, it does not expose the heading of
the device. Since this information is already provided in the `didUpdate
userLocation` delegate method, we can add it to the `UserLocation`
payload for a trivial cost at runtime.